requirements.txt
# ============================================================================
# FINOX SCRAPER - PRODUCTION DEPENDENCIES
# ============================================================================
# Python 3.11+ required
#
# Install: pip install -r requirements.txt
# ============================================================================

# ============================================================================
# WEB SCRAPING
# ============================================================================
httpx==0.25.2                    # Async HTTP client
beautifulsoup4==4.12.2           # HTML parsing
lxml==5.1.0                      # Fast XML/HTML parser
fake-useragent==1.4.0            # User-agent rotation
tenacity==8.2.3                  # Retry logic with exponential backoff

# ============================================================================
# DATA PROCESSING & VALIDATION
# ============================================================================
pandas==2.1.4                    # Data manipulation
pydantic==2.5.3                  # Data validation and settings
pydantic-settings==2.1.0         # Settings management
PyYAML==6.0.1                    # YAML configuration parsing

# ============================================================================
# DATABASE
# ============================================================================
SQLAlchemy==2.0.25               # ORM and database toolkit
alembic==1.13.1                  # Database migrations
PyMySQL==1.1.0                   # MySQL driver
aiomysql==0.2.0                  # Async MySQL driver
cryptography==41.0.7             # Required for MySQL authentication
motor==3.3.2                     # Async MongoDB driver
pymongo==4.6.1                   # Sync MongoDB driver
gunicorn==21.2.0                 # Production WSGI server

# ============================================================================
# API FRAMEWORK
# ============================================================================
fastapi==0.108.0                 # Modern async web framework
uvicorn[standard]==0.25.0        # ASGI server
python-multipart==0.0.6          # Form data parsing

# ============================================================================
# SCHEDULING
# ============================================================================
APScheduler==3.10.4              # Job scheduling
pytz==2023.3.post1               # Timezone support

# ============================================================================
# MONITORING & LOGGING
# ============================================================================
prometheus-client==0.19.0        # Prometheus metrics
loguru==0.7.2                    # Advanced logging

# ============================================================================
# UTILITIES
# ============================================================================
python-dotenv==1.0.0             # Environment variable management
python-dateutil==2.8.2           # Date parsing utilities
requests==2.31.0                 # HTTP library (fallback)

# ============================================================================
# DEVELOPMENT DEPENDENCIES (Optional - install separately)
# ============================================================================
# pytest==7.4.3                  # Testing framework
# pytest-asyncio==0.21.1         # Async test support
# pytest-cov==4.1.0              # Coverage reporting
# black==23.12.1                 # Code formatting
# flake8==7.0.0                  # Linting
# mypy==1.8.0                    # Type checking
# ipython==8.19.0                # Enhanced REPL
